Move the disconnect logic to a seprate folder
Since its unrelated to routing
