+++ src/NServiceBus.Core/EndpointNotifications.cs
@@ -20,7 +20,7 @@ namespace NServiceBus
 
         internal void InvokeSafeToDisconnect(Dictionary<string, string> headers)
         {
-            disconnectEventList.Publish(new SafeDisconnect(new Dictionary<string, string>(headers)));
+            disconnectEventList.Publish(new EndpointSafeToDisconnected(new Dictionary<string, string>(headers)));
         }
 
         void IDisposable.Dispose()
